+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - Occurrence notification subscription always being shown as 0
 - Total capacity count in event and event groups list
-- Login routine failures when redirect target is unclear
+- Flaky login when redirect target was unclear or the sign in callback page itself
 
 ## [1.5.5] - 2021-04-26